refactor(agents): extract terminal-runner projection from lifecycle_ops
Third and final carve of the lifecycle_ops.rs god-file (the most fix-churned file
in the codebase), rebased onto main after the transport-proxy (#9256) and
candidate-adoption (#9260) carves merged. Extracts the terminal runner-result
projection cluster — the read-back half of the Lab handoff — into
lifecycle_runner_projection.rs: reconcile_runner_job_snapshot,
project_terminal_runner_result, project_terminal_runner_job_snapshot,
terminal_runner_lifecycle_event, preserve_terminal_runner_identity,
merge_live_provider_handles, validate_runner_job_snapshot/validate_terminal_child_identity,
and aggregate_projection_plan(_from_outcomes).

Behavior-preserving pure move following the agent_task_lifecycle module pattern
(mod X; pub use X::*). New module opens with use super::* and is re-exported through
mod.rs, so external agent_task_lifecycle:: paths (including the already-merged
transport-proxy module's calls to reconcile_runner_job_snapshot and
terminal_runner_lifecycle_event) resolve unchanged. Widened aggregate_projection_plan
and record_runner_job_terminal_metadata to pub(crate).

lifecycle_ops.rs: 3054 -> 2629 LOC (3708 -> 2629 across all three carves).
