+++ crates/homeboy-agents/src/agent_task_lifecycle/mod.rs
@@ -40,6 +40,7 @@ mod health;
 mod lifecycle_candidate_adoption;
 mod lifecycle_ops;
 mod lifecycle_record_ops;
+mod lifecycle_runner_projection;
 mod lifecycle_transport_proxy;
 mod records;
 pub mod runner_continuation;
@@ -52,6 +53,7 @@ pub use homeboy_core::controller_runtime::ControllerRuntimePruneResult;
 pub use lifecycle_candidate_adoption::*;
 pub use lifecycle_ops::*;
 pub use lifecycle_record_ops::cook_attempt_run_id;
+pub use lifecycle_runner_projection::*;
 pub use lifecycle_transport_proxy::*;
 pub use records::*;
 #[cfg(any(test, feature = "test-support"))]
